+++ e2e_tests/basic_session/conftest.py
@@ -6,28 +6,11 @@ import pytest
 
 from llama_deploy import (
     ControlPlaneConfig,
-    SimpleMessageQueueConfig,
     WorkflowServiceConfig,
-    deploy_core,
     deploy_workflow,
 )
 
 from .workflow import SessionWorkflow
-
-
-def run_async_core():
-    asyncio.run(deploy_core(ControlPlaneConfig(), SimpleMessageQueueConfig()))
-
-
-@pytest.fixture(scope="package")
-def core():
-    p = multiprocessing.Process(target=run_async_core)
-    p.start()
-    time.sleep(5)
-
-    yield
-
-    p.kill()
 
 
 def run_async_workflow():
