+++ e2e_tests/conftest.py
@@ -1,0 +1,23 @@
+import asyncio
+import multiprocessing
+import time
+
+import pytest
+
+from llama_deploy import ControlPlaneConfig, SimpleMessageQueueConfig
+from llama_deploy.deploy import deploy_core
+
+
+def run_async_core():
+    asyncio.run(deploy_core(ControlPlaneConfig(), SimpleMessageQueueConfig()))
+
+
+@pytest.fixture(scope="package")
+def core():
+    p = multiprocessing.Process(target=run_async_core)
+    p.start()
+    time.sleep(5)
+
+    yield
+
+    p.kill()
